Execuitve Biohacking for Travelers/requirements.txt
# cTrader Trading Bot Dependencies
# Core web framework
flask>=2.3.0

# Twisted framework for async networking
twisted>=22.10.0

# cTrader Open API client
ctrader-open-api>=0.9.2

# Protocol Buffers (compatible with ctrader-open-api)
protobuf==3.20.1

# HTTP requests library
requests>=2.28.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0
xgboost

# Data processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv
pymongo
ta

# Interactive input with timeout
inputimeout>=1.0.4

# Additional utilities (standard library, but listed for clarity)
# datetime - standard library
# json - standard library
# time - standard library
# os - standard library
# sys - standard library
# subprocess - standard library
# threading - standard library
# calendar - standard library
# webbrowser - standard library
# importlib - standard library
# logging - standard library
# copy - standard library
